src\block\block_open.c(320): warning C4267: '=': conversion from 'size_t' to 'uint32_t', possible loss of data

+++ src/block/block_open.c
@@ -317,7 +317,7 @@ __wt_desc_write(WT_SESSION_IMPL *session, WT_FH *fh,
 	desc->minorv = WT_BLOCK_MINOR_VERSION;
 	desc->checksum = 0;
 	desc->allocsize = allocsize;
-	desc->metadata_len = metadata_len;
+	desc->metadata_len = (uint32_t)metadata_len;
 	__wt_block_desc_byteswap(desc);
 	desc->checksum = __wt_checksum(desc, allocsize);
 #ifdef WORDS_BIGENDIAN
